Modificacion Nombres segun PEP8
